requirements.txt
numpy
keras
matplotlib
scikit-learn
Flask
Pillow